requirements.txt
Django==1.10.4
django-modeltranslation==0.12
django-simple-captcha==0.5.3
olefile==0.44
Pillow==4.0.0
six==1.10.0
django-rosetta
